url-shorterner/requirements.txt
unique-id-generator==0.0.2
django~=3.1
djangorestframework==3.12.2
django-cors-headers
psycopg2-binary==2.8.5